% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Competitive_exclusion.R
\name{compex}
\alias{compex}
\title{Competitive exclusion (logistic growth) of two species}
\usage{
compex(state = c(X1 = NULL, X2 = NULL), parameters = c(r1 = NULL, r2 =
  NULL, K1 = NULL, K2 = NULL, A12 = NULL, A21 = NULL), times,
  method = NULL)
}
\arguments{
\item{state}{initial state of the populations, X1 and X2}

\item{parameters}{vector of named values, r(1,2) (pop growth rates), K(1,2) (carrying capacities of pops) and A(12, 21) competitive advantages of each species.}

\item{times}{vector of time period (or generations) for logistic equation to be solved over, best generated using the seq() function.}

\item{method}{see ?deSolve::ode for methods and algorithms with which to solve differential equations}
}
\description{
Modifying the logistic regression to two species, adding an extra parameter, alpha
which describes the competitive ability of the population.
}
\examples{
# library(ggplot2); library(dplyr)

# lots of parameters
parameters <- c(r1 = 0.15,
                r2 = 0.1,
                K1 = 100,
                K2 = 100,
                A12 = 0.2,
                A21 = 0.3)
# states of the initial populations
state = c(X1 = 1,
          X2 = 1)
# time in generations
times <- seq(0, 100, by = 0.01)

# run the model
compex(state = state, parameters = parameters, times = times) \%>\%
melt.data.table(measure.vars = c("X1", "X2")) \%>\% 
ggplot(aes(x = time, y = value)) + geom_line(aes(colour = variable))
}
