% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Susceptibe_Infectious.R
\name{SIS}
\alias{SIS}
\title{SIS Model (Susceptible, Infectious, Susceptible)}
\usage{
SIS(state = c(S = NULL, I = NULL), parameters = c(B = NULL, d = NULL, g
  = NULL, beta = NULL), times, method = NULL)
}
\arguments{
\item{state}{initial state of the population classes, susceptible, S, and Infectious, I.}

\item{parameters}{vector of named values, B - birth rate per individual, d is the natural death rate, g is the recovery rate and beta is the contact rate.}

\item{times}{vector of time period (or generations) for SIS equation to be solved over, best generated using the seq() function.}

\item{method}{see ?deSolve::ode for methods and algorithms with which to solve differential equations}
}
\description{
The simplest model from a class of 'SIR' models.
}
\examples{
# library(ggplot2); library(dplyr)

# parameters
parameters <- c(B = 0.21,
                d = 0.2,
                g = 1,
                beta = 3)
                
# states of the initial populations
state = c(S = 1,
          I = 1)
# time in generations
times <- seq(0, 100, by = 0.01)

# run the model
SIS(state = state, parameters = parameters, times = times) \%>\%
melt.data.table(measure.vars = c("S", "I")) \%>\% 
ggplot(aes(x = time, y = value)) + geom_line(aes(colour = variable))
}
