% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logistic_growth.R
\name{loggr}
\alias{loggr}
\title{Logistic growth

Simple differential equation solution to a population undergoing logistic growth. 
It is essentially a wrapper for the deSolve package function ode().}
\usage{
loggr(state = c(X = NULL), parameters = c(r = NULL, K = NULL), times,
  method = NULL)
}
\description{
Logistic growth

Simple differential equation solution to a population undergoing logistic growth. 
It is essentially a wrapper for the deSolve package function ode().
}
\examples{
# library(ggplot2)

# set up the parameters r, population growth rate
# and K, the carrying capacity of the population
parameters <- c(r = 0.1,
                K = 50)
# set the initial parameter for X (pop size)
state <- c(X = 1)
# time in generations
times <- seq(0, 100, by = 0.01)

# run the model
loggr(state = state, parameters = parameters, times = times) \%>\%
   ggplot(aes(x = time, y = X)) + geom_line()
}
